requirements.txt
Django==5.0.1
mysqlclient==2.2.7
djangorestframework==3.16.1
django-cors-headers==4.3.1
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-extensions==4.1
django-filter==25.1
whitenoise==6.6.0
pillow==11.3.0
reportlab==4.4.0
requests==2.31.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psutil==5.9.5
django-redis==6.0.0